functions/scores/requirements.txt
# Ethioware – Scores ingestion (GCS finalize → BigQuery)
# Python 3.10+ (Cloud Functions runtime)

google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
pandas>=2.0.0
cloudevents==1.11.0
